Bug fixes and optimizations
Fixed the bug where it was constantly moving the 2nd+ boid back to the firing position every frame. This happened because it did it every time it had waited MORE than 1.5 seconds, without checking if it has already moved it.

+++ OpenGL Graphics Programming/Math.h
@@ -64,7 +64,7 @@ namespace Math
 		return (reversedOutput) ? (newMax - portion) : (portion + newMin);
 	}
 
-	const unsigned int Box2DScale = 100;
+	const float Box2DScale = 100.0f;
 
 	inline glm::vec2 Box2DtoVec2(b2Vec2 _vec)
 	{
